New kas_core::core::widget_id module for WidgetId
